Kernel: Introduce the EFIPrekernel
The EFIPrekernel currently only works on RISC-V, as only supporting that
architecture for now was easier.
This commit also doesn't add support for parsing the (U)EFI-provided
memory map and cmdline. The kernel image path is hardcoded for now as
well.
This is currently a separate Prekernel until we refactor the x86
multiboot Prekernel so we can merge these two Prekernels.

This Prekernel gets converted to a PE32+ EFI image via the
PrekernelPEImageGenerator.